Implement drop down in RoundedToolbar
The RoundedToolItem supports the style SWT.DROP_DOWN, but setting the
style has no effect and there is no obvious way to supply any menu to
the item even though there are some traces in the code the style should
actually do something, it therefore feels like the implementation is
incomplete.

This now adds a RoundedToolItem#setMenuCreator similar to what JFace
supports for actions with a menu.