[grid] Add basic fixed column support
Currently Nebula Grid does not offer a feature know as "fixed columns"
where one can set some columns to be displayed regardless of the
scrolling.

This now adds a very basic first support for such columns, even though
it does not work for all grid features it allows to fix a column in a
simple grid.