Fix cell text is invisible on selected lines with fixed colums
Currently when one has a fixed column and selects a line then scrolls
the cell text becomes invisible. This is because we effectively paint
twice and the empty cell renderer thinks we have no more columns and
overwrite in some cases.

THis now passes a flag to the call to skip the empty painting as we have
done this anyways with the previous call.