Workaround for Race Condition on Creating Browserfunctions for Edge #662
On the edge browser engine it can happen that the browser functions
registered by the rich text editor are not yet available when the
document is parsed. By setting the template url after the browser
functions are registered and sleeping the thread a bit the race
condition went away.

Addresses: https://github.com/EclipseNebula/nebula/issues/662
See also:
https://github.com/eclipse-platform/eclipse.platform.swt/issues/2449